+++ script/install.sh
@@ -1,1 +1,4 @@
 npm install && composer install;
+mkdir -p public/css public/js;
+touch public/css/main.css;
+touch public/js/main.css;